+++ backend/emm-sale/src/test/java/com/emmsale/notification/application/NotificationEventListenerTest.java
@@ -7,8 +7,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.emmsale.comment.domain.Comment;
-import com.emmsale.comment.event.UpdateNotificationEvent;
+import com.emmsale.event_publisher.UpdateNotificationEvent;
 import com.emmsale.helper.ServiceIntegrationTestHelper;
 import com.emmsale.notification.domain.UpdateNotification;
 import com.emmsale.notification.domain.UpdateNotificationRepository;
